rename files to camelCase
