+++ sandpack-react/src/hooks/useSandpackTheme.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { SandpackThemeContext } from '../contexts/theme-context';
+import { SandpackThemeContext } from '../contexts/themeContext';
 
 export const useSandpackTheme = () => {
   const { theme, id } = React.useContext(SandpackThemeContext);
