+++ sandpack-react/src/index.ts
@@ -23,8 +23,8 @@ export * from './hooks/useSandpackNavigation';
 export * from './hooks/useSandpackTheme';
 export * from './hooks/useTranspiledCode';
 
-export * from './contexts/sandpack-context';
-export * from './contexts/theme-context';
+export * from './contexts/sandpackContext';
+export * from './contexts/themeContext';
 export * from './themes';
 export * from './types';
