+++ sandpack-react/src/utils/stringUtils.ts
